==> 1.Capstone Decision.ipynb <==
# Capstone Project Decision

# Part_1: Capstone Ideas and Decision Making

## Requirements

1. Identifies which of the three proposals you outlined in your lightning talk you have chosen
2. Articulates the main goal of your project (your problem statement)
3. Outlines your proposed methods and models
4. Defines the risks & assumptions of your data
5. Revises initial goals & success criteria, as needed
6. Documents your data source
7. Performs & summarizes preliminary EDA of your data

### Project Ideas

**Option A**
- Natural Language Processing/Inference

***Task*** - Drawing inference from health research publications Using Natural Process Inference 
Corpus semantics – Applying corpus methods to what sentences mean rather than what type of sentence they are (sentiments)

***Data*** -
From Stanford Natural Language Inference Corpus
A collection of 570k labeled human-written English sentence pairs

***Expected Outcome*** -
For each fullVisitorId in the test set, you must predict the natural log of their total revenue in PredictedLogRevenue. 


**Option B**
-  Kaggle Competition

**Task** -
Google Analytics Customer Revenue Prediction -
Predict how much Gstore customers will spend

**Data** -
Given on Kaggle, there are some json scripts in the dataset. Data-size for training = 1,468,195 KB; Data-size for test = 1,315,279 KB; Data format = .csv

**Expected Outcome** -
For each fullVisitorId in the test set, you must predict the natural log of their total revenue in PredictedLogRevenue. 

**Option C**
- Movie Ratings

**Task** -
Predicting movie ratings for users

**Data** -
iIMDb Datasets

**Expected Outcome** -
Being able to predict which moving a person might like of another
Motivation: I have waisted enough time watching bad movies and would like to know once I should be investing my time in and which not to.




### Lasso Modeling
- After applying Laso to my three ideas, Option A and C where zeroed out which leaves us with **option B - Kaggle competition**.

I will say this is a safe choice because the data sets and questions are given, although not an easy option, based on past experience with kaggle competitions, plus there is a potential financial /leader-board 'fame' incentive that comes with this project.


## Deciding on Option B - Kaggle competition.

### Problem Statement
**Google Analytics Customer Revenue Prediction**

   - Predict how much GStore customers will spend

### Proposed Methods and Models

> I will be trying out different unsupervised and supervised or a combination of both type of models to find out which one.

### Defines the risks & assumptions of your data
**Risks** involved will be minimal. some of which includes having sufficient time to try out all of the intended models and also have insufficient memory to run some of the models 

Also this might be too aggressive given the nature/complexity of the dataset. 

### Revise initial goals & success criteria, as needed
> My goals remain the same as the criteria have been layout by kaggle competition. However, I may want to predict whether a site visit led to a transaction and if yes, go ahead to predict amount customers will spend per transaction.

### Data Source
> Data - Given on Kaggle. 
Data-size for training = 1,468,195 KB; 
Data-size for test = 1,315,279 KB; 
Data format = .csv; There are some json scripts in the dataset


### Examining Data Set

import pandas as pd
import numpy as np

df = pd.read_csv("../kaggle_data/train.csv")

df.head()

df.info()

df.describe()
