==> 4.Captone_Project -Part_4-EDA_and_Base Model.ipynb <==
# Capstone Project - Google Predictive Analytics

# Part_4: EDA and Base Model

### Extensive EDA of our data and base model

import pandas as pd
import numpy as np

df = pd.read_csv("../capstone_data/kaggle_data/train.csv")

df.head()

df.info()

df.describe()

df.columns

#### Exploring each column to understand contents

df['channelGrouping'].value_counts() # Can get_dummies with this variable

df['channelGrouping'].isnull().sum()

from datetime import datetime
df['date'].value_counts()[:5] # stored as int. possibly changing to date timestamp
df['date'] = pd.to_datetime(df['date'])
df['date'].value_counts()[:5]


df['date'].isnull().sum()

# stored as a json list. 
#most of the data is not available in demo dataset and may not be relevant in model. 
#Might end up not including in data
df['device'][3]

import json

device = pd.DataFrame([json.loads(each) for each in df['device']])

device.head(2)

device.isnull().sum()

for each in device:
    print(device[each].value_counts()[:3])

device.columns

# Droping all "not availble in demo dataset" 
device = device.drop(['browserSize', 'browserVersion', 'flashVersion', 'language', 'mobileDeviceBranding',
       'mobileDeviceInfo', 'mobileDeviceMarketingName', 'mobileDeviceModel',
       'mobileInputSelector', 'operatingSystemVersion',
       'screenColors', 'screenResolution' ], axis=1)

device.head(2)

device['isMobile'].value_counts()

device['deviceCategory'].value_counts()

device['isMobile'].isnull().sum()

device['deviceCategory'].isnull().sum()

device[(device['isMobile'] == False) & (device['deviceCategory'] == 'mobile')].groupby(['operatingSystem', 'browser']).count()[['deviceCategory']]

device['operatingSystem'].value_counts()


df.shape

device.shape

# joining the device sub-columns to df while deleting device. 
# also Dropping isMobile because information is contained in deviceCategory
df = pd.concat([df, device], axis=1).drop(['device','isMobile'], axis=1)

df.head()

df['fullVisitorId'].head()

#df['fullVisitorId'].astype('int')

# unable to cast as an int due to int memory overflow error
# casting as a float instead. 
# May not to do much with this since it's an id and they have very little impact on models
df['fullVisitorId'].astype('float').head()

df['geoNetwork'][0]

# stored as a json script. 
# might draw correlation from continent and country, hence get_dummies
geoNetwork = pd.DataFrame([json.loads(each) for each in df['geoNetwork']])

geoNetwork.head()

geoNetwork.city.value_counts().head()

geoNetwork.cityId.value_counts()

geoNetwork.latitude.value_counts()

geoNetwork.longitude.value_counts()

geoNetwork.networkLocation.value_counts()

# will drop because over 70,000 is not available nor set. The values available are in city.
geoNetwork.metro.value_counts()[:5]

geoNetwork.networkDomain.value_counts()[:5]

len(geoNetwork.city) - sum(geoNetwork.city == geoNetwork.metro)

geoNetwork = geoNetwork.drop(['networkLocation', 'longitude', 'latitude', 'cityId', 'metro', 'country', 'region', 'subContinent'], axis=1)

geoNetwork.head()

geoNetwork.isnull().sum()

# adding geoNetwork sub-columns to df
df = pd.concat([df, geoNetwork], axis=1).drop('geoNetwork', axis=1)

df.head(3)

# sessionID has fullVisitorId plus _and session id. 
# might hold same relevance as fullVisitorId 
df['sessionId'][0]

df['fullVisitorId'][0]

df['sessionId'][0].split('_')[-1]

df['sessionId'] = df['sessionId'].astype("float")
df['fullVisitorId'] = df['fullVisitorId'].astype("float")

# variable has same value for all observations. might not include column in model
df['socialEngagementType'].value_counts()
df.drop('socialEngagementType', axis=1, inplace=True)

# stored as a json string- showing online presence/status to site.
# might use newVisits number in model if different for observations 
# also might contain our y-label. Will explore in details further down.
df['totals'][0]

json.loads(df['trafficSource'][2])

# this is a json script that contain values that may be valuable to our model. 
#might make use of some of the keys in the dictionary by getting dummies on the useful ones.
trafficSource = pd.DataFrame([json.loads(each) for each in df['trafficSource']])

trafficSource.head()

trafficSource.isnull().sum()

trafficSource.adContent.value_counts()

# Has the same value for all records. will drop.
trafficSource.adwordsClickInfo.values

# We could assume that for "not set" entries, no campaign was made. Customers went to the site on their own.
# Which is is more than 80% of visits.
trafficSource.campaign.value_counts()[:2]

# will drop campaignCode since its all null with exception of 1
trafficSource.campaignCode.value_counts()

trafficSource.isTrueDirect.notnull().value_counts()

trafficSource.keyword.value_counts()[:5] # Might fill nan's with -999

trafficSource.medium.value_counts() # will drop. ChannelGroupping has similar and more precise information.

trafficSource.referralPath.value_counts()[:5]

trafficSource.source.value_counts()[:5]

pd.DataFrame(df['trafficSource'].value_counts()).head()

trafficSource.columns

trafficSource = trafficSource.drop(['adwordsClickInfo', 'campaignCode'], axis=1)

trafficSource = trafficSource.fillna(-999)

trafficSource.head()

df = pd.concat([df, trafficSource], axis=1).drop(['trafficSource', 'medium'], axis=1)

df.head()

# stored as ints, will be used in correlation and modela
df['visitId'].value_counts().sort_values(ascending=False)[:5]

# stored as ints, could be used in correlation and model
df['visitNumber'].value_counts().sort_values(ascending=False)[:5]

# stored as ints, values and counts are same as visitId. 
# may be redundant since its a duplicate otherwise, could be used in correlation and model
df['visitStartTime'].value_counts().sort_values(ascending=False)[:5]

import json
# This is my y-label for this data set. What I will be predicting.

df["totals"].head()

totals = pd.DataFrame([json.loads(each) for each in df['totals']])

totals.head()

totals.isnull().sum()

# Fill all Nan's in total df with seems it seems no value was 0
totals = totals.fillna(0)

totals.info()

# They are all numerical values.
totals.describe()

totals['visits'].value_counts() # Should be removed from df since it has the same value for each observation.

totals = totals.drop(['visits'], axis=1)

totals.columns

totals['bounces'] = totals.bounces.astype('int')
totals['hits'] = totals['hits'].astype('int')
totals['newVisits'] = totals['newVisits'].astype('int')
totals['pageviews'] = totals['pageviews'].astype('int')

totals['transactionRevenue'].shape

#  from isnull(), we realize there are 892137 Nan's which means that 
# there were 11516 visits that resulted into a purchases.
# NaN's will be filled in with 0
903653 - 892137

# for the values that are not null, we will divide by 1000000 
# because the values are to the 1,000,000th unit.
totals[totals['transactionRevenue'].notnull()]['transactionRevenue'].value_counts().sort_values(ascending=True)[:3]

totals['transactionRevenue'] = totals['transactionRevenue'].map(lambda x: float(x)/1000000)
totals["transactionRevenue"].sort_values().head()

min_total_transactionRevenue = totals['transactionRevenue'].sort_values(ascending=True)
min_total_transactionRevenue.head()

max_total_transactionRevenue = totals['transactionRevenue'].sort_values(ascending=False)
max_total_transactionRevenue.head()

totals.head()

df = pd.concat([df, totals], axis=1).drop('totals', axis=1)

df.head()

df.isnull().sum()

# Export to csv
df.to_csv('../capstone_data/kaggle_data/df_combined_tables.csv', index=False)

### Creating a basic model using only the values of total.

totals.head(3)

totals.isnull().sum()

totals.columns

totals.info()

# from sklearn.preprocessing import normalize

y = totals['transactionRevenue']
X = totals.drop(['transactionRevenue'], axis=1)

y.head()

#### train_test_split

from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split, cross_val_score

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

X_train.corrwith(y_train).sort_values()

lr = LinearRegression()
lr.fit(X_train, y_train)

lr.score(X_train, y_train), lr.score(X_test, y_test)

from sklearn.metrics import mean_squared_error

print(mean_squared_error(y_train, lr.predict(X_train)) ** 0.5)
print (mean_squared_error(y_test, lr.predict(X_test)) ** 0.5)  

**This large value so far away from zero shows that our model has a long way to improve on.**
