==> 6.Captone_Project -Part_6-Modeling-PCA.ipynb <==
# Capstone Project - Google Predictive Analytics

## Part_5: Modeling 

### Performs some more EDA and Modeling

import pandas as pd
import numpy as np

import matplotlib
import matplotlib.pyplot as plt
import seaborn as sns
plt.style.use('fivethirtyeight')

%config InlineBackend.figure_format = 'retina'
sns.set_style('whitegrid')
plt.style.use('fivethirtyeight')
%matplotlib inline

df = pd.read_csv("../capstone_data/kaggle_data/df_combined_tables.csv")

df.head()

# going ahead to drop city because it shows quite a bit of 'not available in demo dataset' 
# but the continent is present which should be sufficient. Also isTrueDirect, keyword and referralPath would be dropped 
# because of too many nan values and some of the content with other columns could be used to explain these columns. 
df = df.drop(['city', 'isTrueDirect', 'keyword', "referralPath", 'campaign', 'date'], axis=1)

# I will be dropping some more categorical varibles due to memory errors when getting dummies. 
# We are working with over 900,000 records. 
df = df.drop(['operatingSystem', "networkDomain"], axis=1)

df.info()

df.head()

df.shape

df_categorical = list(df.select_dtypes(include=['object']).columns)

df[df_categorical].head(10)

df_dummies = pd.get_dummies(df[df_categorical])
df_dummies.head()

df_dummies.shape

numerical_features = list(df._get_numeric_data().columns)

df[numerical_features].head()

df[numerical_features].shape

##### Combining categorical and numerical features

df_2 = pd.concat([df_dummies, df[numerical_features]], axis=1)

# Export to csv
df_2.to_csv('../capstone_data/kaggle_data/df_dummies_tables.csv', index=False)

X_df = df_2.drop('transactionRevenue', axis=1)

X_df.head()

y= df.transactionRevenue

y.head()

#### Train_test_split and model with Linear Regression

y.head()

from sklearn.linear_model import RidgeCV, LassoCV, ElasticNetCV, LinearRegression
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler, PolynomialFeatures

ss = StandardScaler()
ss.fit(X_df)

X_s = ss.transform(X_df)

X_s

X_train, X_test, y_train, y_test = train_test_split(X_s, y, random_state=42)

# X_train.corrwith(y_train).sort_values()

lr = LinearRegression()
lr.fit(X_train, y_train)

lr.score(X_train, y_train), lr.score(X_test, y_test)

from sklearn.metrics import mean_squared_error

print(mean_squared_error(y_train, lr.predict(X_train)) ** 0.5)
print (mean_squared_error(y_test, lr.predict(X_test)) ** 0.5)  

# pipe = Pipeline([
#     #('pf', PolynomialFeatures()),
#     ('ss', StandardScaler()),
#     ('lr', LinearRegression())
# ])

# pipe.fit(X_train, y_train)

# print(pipe.score(X_train, y_train))
# print(pipe.score(X_test, y_test))

##### Using Ridge and lasso regressing

rcv = RidgeCV(alphas=np.logspace(0, 5,100), cv = 3)

rcv.fit(X_train, y_train)

rcv.score(X_train, y_train), rcv.score(X_test, y_test)

print(mean_squared_error(y_train, rcv.predict(X_train)) ** 0.5)
print (mean_squared_error(y_test, rcv.predict(X_test)) ** 0.5) 

lcv = LassoCV(cv=3)

lcv.fit(X_train, y_train)

lcv.score(X_train, y_train), lcv.score(X_test, y_test)

from sklearn.metrics import mean_squared_error

print(mean_squared_error(y_train, lcv.predict(X_train)) ** 0.5)
print (mean_squared_error(y_test, lcv.predict(X_test)) ** 0.5)  

variables = X_df.columns
coef_values = lcv.coef_

variables_dictionary = {
    'variables': variables,
    'coefficient': coef_values
}

pd.DataFrame(variables_dictionary).set_index('variables').sort_values('coefficient', ascending=False).head(20)
