==> 7-Captone_Project -Modeling-comparison.ipynb <==
# Capstone Project - Google Predictive Analytics

## Part_7: Modeling_Comparison

### Comparing models listed below and their perfomances
- LinearRegression
- DecisionTreeRegressor
- ExtraTreesRegressor
- RandomForestRegressor
- AdaBoostRegressor
- GradientBoostingRegressor
- Simple Vector Machine 

import pandas as pd
import numpy as np

import matplotlib
import matplotlib.pyplot as plt
import seaborn as sns
plt.style.use('fivethirtyeight')

%config InlineBackend.figure_format = 'retina'
sns.set_style('whitegrid')
plt.style.use('fivethirtyeight')
%matplotlib inline

df = pd.read_csv('../capstone_data/kaggle_data/df_dummies_tables.csv')

df.shape

#### Train_test_split and model with various Regression

from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler, PolynomialFeatures

from sklearn.linear_model import LinearRegression
from sklearn.neighbors import KNeighborsRegressor
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import ExtraTreesRegressor, RandomForestRegressor, AdaBoostRegressor, GradientBoostingRegressor
from sklearn.svm import SVR

ss = StandardScaler()
ss.fit(df)

X_s = ss.transform(df)

X_s

y = df['transactionRevenue']
y.head()

X_train, X_test, y_train, y_test = train_test_split(X_s, y, random_state=42)

lr = LinearRegression()
dtr = DecisionTreeRegressor()
etr = ExtraTreesRegressor()
rfr = RandomForestRegressor()
abr = AdaBoostRegressor()
gbr = GradientBoostingRegressor()
svr = SVR()

list_of_models = [lr, dtr, etr, rfr, abr, gbr, svr]

for each in list_of_models:
    print(str(each).split('(')[0])
    each.fit(X_train, y_train)
    print('Training score:')
    print(each.score(X_train, y_train))
    print('Testing score:')
    print(each.score(X_test, y_test))
    print()

gbr.fit(X_train, y_train)

gbr.score(X_train, y_train), gbr.score(X_test, y_test)

from sklearn.metrics import mean_squared_error


def str_from_model(model):
    return str(model).split('(')[0]

for each in list_of_models:
    print('>', str_from_model(each))
    print('Training RMSE:')
    print(mean_squared_error(y_train, each.predict(X_train)) ** 0.5)
    print (mean_squared_error(y_test, each.predict(X_test)) ** 0.5)
    print('\n**********\n')

from sklearn.metrics import mean_squared_error

print(mean_squared_error(y_train, gbr.predict(X_train)) ** 0.5)
print (mean_squared_error(y_test, gbr.predict(X_test)) ** 0.5)
